build(FRA-2253) Fetch status check from Sonar and block build if it fails (#338)
Add an additional step to verify the Sonar check.
To verify that it's passed:

send request to SonarCloud (include projectKey and pullRequest to fetch status about exactly this PR)
check projectStatus.status from the response. It's an overall status of the check, so if at least 1 check hasn't passed, the result will be not "OK" (either "FAILED", or some other options, doesn't matter for us)
return 0 from the step if the response if OK, 1 otherwise
Additionally I introduced a parameter blockPullRequestIfSonarFails. If the parameter is false (by default), then don't run this step. In this case we can add this parameter to repositories separately and make sure that it works for each of them.